fix(utils): fix num to double conversion
- fix(booking_manager): check for date being null

+++ lib/utils/color_extension.dart
@@ -20,7 +20,7 @@ extension ColorExtension on Color {
 
       if (currentValue <= 1) continue;
 
-      final colorValue = currentValue.map(inMax: highestValue);
+      final colorValue = currentValue.map(inMax: highestValue).toDouble();
 
       if (!colorMap.containsKey(currentValue)) {
         colorMap.addAll({currentValue: withOpacity(colorValue)});
