+++ lib/widgets/standalone/floating_action_button/animated_floating_button_label.dart
@@ -23,10 +23,12 @@ class AnimatedFloatingButtonLabel extends AnimatedWidget {
       margin: const EdgeInsets.only(right: 18.0),
       padding: const EdgeInsets.symmetric(vertical: 5.0, horizontal: 8.0),
       child: Opacity(
-        opacity: animation.value.map(
-          inMin: (tween.end ?? 0.0) / 2.0,
-          inMax: tween.end ?? 0.0,
-        ),
+        opacity: animation.value
+            .map(
+              inMin: (tween.end ?? 0.0) / 2.0,
+              inMax: tween.end ?? 0.0,
+            )
+            .toDouble(),
         child: label,
       ),
     );
